Report throughput as integer

+++ lm/kenlm_benchmark_main.cc
@@ -65,7 +65,7 @@ template <class Model, class Width> void QueryFromBytes(const Model &model, int 
   std::cout << "Queries: " << completed << std::endl;
   double cpu_per_entry = ((after - loaded) / static_cast<double>(completed));
   std::cout << "CPU_excluding_load: " << (after - loaded) << "\nCPU_per_query: " << cpu_per_entry << std::endl;
-  std::cout << "Throughput: " << (1.0/cpu_per_entry) << std::endl;
+  std::cout << "Throughput: " << (int)(1.0/cpu_per_entry) << std::endl;
   std::cout << "RSSMax: " << util::RSSMax() << std::endl;
 }
